requirements.txt
# torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
opencv-python>=4.5.0
pillow>=8.3.0
tqdm>=4.62.0
pyyaml>=6.0
matplotlib>=3.5.0
tensorboard>=2.8.0
lpips>=0.1.4

# GPU优化相关依赖（可选）
# nvidia-ml-py3>=11.495.46  # GPU监控，获取GPU利用率、温度等信息（可选，某些环境可能不支持）
psutil>=5.8.0            # 系统监控，CPU、内存使用情况 